search_module/requirements.txt
lancedb>=0.4.0
openai>=1.0.0
numpy>=1.20.0
pandas>=1.3.0
python-dotenv>=0.19.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
tenacity>=8.0.0
